requirements.txt
mypy==1.14.1
mypy-extensions==1.0.0
PyYAML==6.0.2
types-PyYAML==6.0.12.20241230
typing_extensions==4.12.2
